refactor(benchmark): move AuditCheckSetup to auditor_setup.go
