+++ token/core/zkatdlog/nogh/v1/benchmark/issue_setup.go
@@ -11,14 +11,11 @@ import (
 	"github.com/hyperledger-labs/fabric-smart-client/platform/view/services/metrics/disabled"
 	tokcommon "github.com/hyperledger-labs/fabric-token-sdk/token/core/common"
 	v1 "github.com/hyperledger-labs/fabric-token-sdk/token/core/zkatdlog/nogh/v1"
-	zkcommon "github.com/hyperledger-labs/fabric-token-sdk/token/core/zkatdlog/nogh/v1/crypto/common"
 	v1driver "github.com/hyperledger-labs/fabric-token-sdk/token/core/zkatdlog/nogh/v1/driver"
-	issue_pkg "github.com/hyperledger-labs/fabric-token-sdk/token/core/zkatdlog/nogh/v1/issue"
 	"github.com/hyperledger-labs/fabric-token-sdk/token/driver"
 	drivermock "github.com/hyperledger-labs/fabric-token-sdk/token/driver/mock"
 	"github.com/hyperledger-labs/fabric-token-sdk/token/services/logging"
 	token2 "github.com/hyperledger-labs/fabric-token-sdk/token/token"
-	"go.opentelemetry.io/otel/trace/noop"
 )
 
 // IssueSetup holds a ready-to-call IssueService together with the invocation
@@ -74,104 +71,5 @@ func NewIssueSetup(conf *SetupConfiguration) (*IssueSetup, error) {
 		TokenType: "ABC",
 		Values:    []uint64{100},
 		Owners:    [][]byte{conf.OwnerIdentity.ID},
-	}, nil
-}
-
-// AuditCheckSetup holds a ready-to-call AuditorService together with a
-// pre-built TokenRequest and TokenRequestMetadata for a single issue action.
-// The request is generated once at setup time from real ZK proof material so
-// that AuditorCheck exercises the actual Pedersen commitment arithmetic on
-// every benchmark iteration.
-type AuditCheckSetup struct {
-	Service  *v1.AuditorService
-	Request  *driver.TokenRequest
-	Metadata *driver.TokenRequestMetadata
-	Anchor   driver.TokenRequestAnchor
-}
-
-// NewAuditCheckSetup constructs an AuditCheckSetup from a SetupConfiguration.
-// It generates a real ZK issue action using the crypto layer directly, then
-// wires an AuditorService with the driver Deserializer so that identity
-// verification matches production behaviour.
-func NewAuditCheckSetup(conf *SetupConfiguration) (*AuditCheckSetup, error) {
-	// Build a real issue action using the crypto layer so that the Pedersen
-	// commitment in the action is genuine and AuditorCheck performs real arithmetic.
-	issuerID := conf.IssuerSigner.ID
-	signer := &zkcommon.WrappedSigningIdentity{
-		Identity: issuerID,
-		Signer:   conf.IssuerSigner.Signer,
-	}
-
-	values := []uint64{100}
-	owners := [][]byte{conf.OwnerIdentity.ID}
-
-	issuer := issue_pkg.NewIssuer("ABC", signer, conf.PP)
-	issueAction, zkMeta, err := issuer.GenerateZKIssue(values, owners)
-	if err != nil {
-		return nil, errors.Wrap(err, "failed generating ZK issue for audit check setup")
-	}
-
-	issueActionRaw, err := issueAction.Serialize()
-	if err != nil {
-		return nil, errors.Wrap(err, "failed serializing issue action for audit check setup")
-	}
-
-	metaBytes, err := zkMeta[0].Serialize()
-	if err != nil {
-		return nil, errors.Wrap(err, "failed serializing token metadata for audit check setup")
-	}
-
-	request := &driver.TokenRequest{
-		Issues: [][]byte{issueActionRaw},
-	}
-
-	// The issuer identity stored in the action (ia.Issuer) equals issuerID.
-	// Setting IssueMetadata.Issuer.Identity to the same value satisfies the
-	// identity-equality check inside audit.GetAuditInfoForIssues.
-	metadata := &driver.TokenRequestMetadata{
-		Issues: []*driver.IssueMetadata{
-			{
-				Issuer: driver.AuditableIdentity{
-					Identity:  issuerID,
-					AuditInfo: conf.IssuerSigner.AuditInfo,
-				},
-				Outputs: []*driver.IssueOutputMetadata{
-					{
-						OutputMetadata: metaBytes,
-						Receivers: []*driver.AuditableIdentity{
-							{
-								Identity:  conf.OwnerIdentity.ID,
-								AuditInfo: conf.OwnerIdentity.AuditInfo,
-							},
-						},
-					},
-				},
-			},
-		},
-	}
-
-	ppm, err := tokcommon.NewPublicParamsManagerFromParams(conf.PP)
-	if err != nil {
-		return nil, errors.Wrap(err, "failed creating public params manager for audit check setup")
-	}
-
-	deserializer, err := v1driver.NewDeserializer(conf.PP)
-	if err != nil {
-		return nil, errors.Wrap(err, "failed creating deserializer for audit check setup")
-	}
-
-	service := v1.NewAuditorService(
-		logging.MustGetLogger(),
-		ppm,
-		deserializer,
-		v1.NewMetrics(&disabled.Provider{}),
-		noop.NewTracerProvider(),
-	)
-
-	return &AuditCheckSetup{
-		Service:  service,
-		Request:  request,
-		Metadata: metadata,
-		Anchor:   "benchmark-anchor",
 	}, nil
 }
